pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pixano"
dynamic = ["version"]
description = 'Data-centric AI building blocks for computer vision applications'
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = "CeCILL-C"
authors = [{name = "Pixano Developers", email = "pixano@cea.fr"}]
keywords = [
  "computer vision",
  "deep learning",
  "machine learning",
  "data visualization",
  "data annotation"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: CeCILL-C Free Software License Agreement (CECILL-C)",
]
dependencies = [
  "duckdb ~= 0.10.1",
  "etils ~= 1.1.0",
  "fastapi ~= 0.103.0",
  "fastapi-pagination ~= 0.12.0",
  "httpx ~= 0.25.0",
  "imageio ~= 2.25.0",
  "importlib-resources ~= 5.12.0",
  "ipywidgets ~= 8.0.0",
  "jinja2 ~= 3.1.2",
  "lancedb ~= 0.6.4",
  "numpy < 2.0.0",
  "onnx ~= 1.13.0",
  "onnxruntime ~= 1.15.0",
  "opencv-python ~= 4.7.0",
  "pandas ~= 1.5.0",
  "Pillow ~= 9.4.0",
  "pre-commit ~= 3.6.0",
  "pyarrow ~= 12.0.0",
  "pycocotools ~= 2.0.0",
  "pydantic ~= 2.4.0",
  "pydantic-settings ~= 2.0.0",
  "s3path ~= 0.5.0",
  "setuptools ~= 65.6.0",
  "shortuuid ~= 1.0.0",
  "tqdm ~= 4.64.0",
  "uvicorn ~= 0.20.0",
]

[project.optional-dependencies]
documentation = [
  "mkdocs-material ~= 9.4.0",
  "mkdocstrings-python ~= 1.7.0",
  "mkdocs-gen-files ~= 0.5.0",
  "mkdocs-literate-nav ~= 0.6.0",
  "mkdocs-redirects ~= 1.2.1",
]
test = [
  "mediapy",
  "pytest",
  "pytest-cov",
  "pytest-mock",
  "ffmpeg-python~=0.2.0",
]

[project.urls]
Documentation = "https://github.com/pixano/pixano#readme"
Issues = "https://github.com/pixano/pixano/issues"
Source = "https://github.com/pixano/pixano"

[project.scripts]
pixano = "pixano.app.serve:main"

[tool.hatch.build]
include = ["/pixano"]
exclude = ["__pycache__"]
artifacts = ["/pixano/app/dist"]

[tool.hatch.version]
path = "pixano/__version__.py"

[tool.ruff]
line-length = 119
extend-exclude = ["ui"]
extend-include = ["notebooks/**/*.ipynb"]

[tool.ruff.lint]
ignore=["D205", "C901", "D100", "D104"]
select = ["C", "E", "F", "I", "W", "D"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "C408", "C901", "F401"]
"docs/*" = ["D", "F401"]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["pixano"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
